+++ src/platform/routes/store.route.js
@@ -1,26 +1,26 @@
 'use strict';
-const authenticate = require('../../config/authenticate');
-const PlatformService = require('./../service/platform-service');
+const StoreService = require('../service/store.service');
 
 class StoreController {
     store(req, res) {
-        let ps = new PlatformService(req.user);
-        ps.getApplications(applications => {
+        let ss = new StoreService(req.user);
+        ss.getApplications(applications => {
             res.render('pages/store',
                 {applications: applications, layout: 'layouts/layout', user: req.user});
         });
     }
 
     storeApp(req, res) {
-        let ps = new PlatformService(req.user);
-        ps.getApplicationById(req.params.id, function (application) {
+        let ss = new StoreService(req.user);
+        ss.getApplicationById(req.params.id, function (application) {
             res.render('pages/store-app',
                 {application: application, layout: 'layouts/layout', user: req.user});
         });
     }
 
     storeAppInstall(platform, req, res) {
-        platform.installApplication(req.params.id, req.body.sample).then(function () {
+        let ss = new StoreService(req.user);
+        ss.installApplication(req.params.id, req.body.sample, platform).then(function () {
             res.send({});
         });
     }
